# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DJD Agent Score API",
    "description": "Reputation scoring for autonomous AI agent wallets on Base. Scores are derived from on-chain USDC behavioral data and optional self-registration. Paid endpoints require an x402 USDC micropayment on Base; free endpoints require no authentication. Interactive docs available at /docs.",
    "version": "2.0.0",
    "contact": { "email": "feedback@djdagentscore.dev" },
    "license": { "name": "MIT", "url": "https://github.com/jacobsd32-cpu/djdagentscore/blob/main/LICENSE" }
  },
  "servers": [
    { "url": "https://djdagentscore.dev", "description": "Production" }
  ],
...
  "paths": {
    "/v1/score/basic": {
      "get": {
        "tags": ["score"],
        "summary": "Basic score lookup",
        "description": "Returns composite score, tier, confidence, recommendation, and model version. Free tier: 10 requests/day per IP. No payment required.",
        "operationId": "getBasicScore",
        "parameters": [
          { "$ref": "#/components/parameters/wallet" }
        ],
        "responses": {
          "200": {
            "description": "Score returned (may include stale:true if RPC was unavailable)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BasicScoreResponse" },
                "example": {
                  "wallet": "0x3e4ef1f774857c69e33dddc471e110c7ac7bb528",
                  "score": 13,
                  "tier": "Unverified",
                  "confidence": 0.1,
                  "recommendation": "insufficient_history",
                  "modelVersion": "1.0.0",
                  "lastUpdated": "2026-02-21T20:00:00.000Z",
                  "computedAt": "2026-02-21T20:00:00.000Z",
                  "scoreFreshness": 0.85
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "429": { "$ref": "#/components/responses/FreeTierExhausted" }
        }
      }
    },
    "/v1/score/full": {
      "get": {
        "tags": ["score"],
        "summary": "Full score with dimension breakdown",
        "description": "Returns the composite score plus all four dimension scores (reliability, viability, identity, capability), integrity flags, data availability assessment, and improvement path. **Requires x402 payment of $0.10 USDC on Base.**",
        "operationId": "getFullScore",
        "parameters": [
          { "$ref": "#/components/parameters/wallet" }
        ],
        "security": [{ "x402": [] }],
        "x-x402": { "price": "$0.10", "network": "base" },
        "responses": {
          "200": {
            "description": "Full score returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FullScoreResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/v1/score/refresh": {
      "post": {
        "tags": ["score"],
        "summary": "Force live recalculation",
        "description": "Bypasses the cache and triggers a fresh on-chain scan. Returns the full score response. **Requires x402 payment of $0.25 USDC on Base.** Prefer POST; GET is accepted for backward compatibility but deprecated.",
        "operationId": "refreshScore",
        "parameters": [
          { "$ref": "#/components/parameters/wallet" }
        ],
        "security": [{ "x402": [] }],
        "x-x402": { "price": "$0.25", "network": "base" },
        "responses": {
          "200": {
            "description": "Freshly computed full score",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FullScoreResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      },
      "get": {
        "tags": ["score"],
        "summary": "Force live recalculation (deprecated — use POST)",
        "description": "**Deprecated.** Use `POST /v1/score/refresh` instead. This GET alias is kept for backward compatibility only.",
        "operationId": "refreshScoreDeprecated",
        "deprecated": true,
        "parameters": [
          { "$ref": "#/components/parameters/wallet" }
        ],
        "security": [{ "x402": [] }],
        "x-x402": { "price": "$0.25", "network": "base" },
        "responses": {
          "200": {
            "description": "Freshly computed full score",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FullScoreResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/v1/agent/register": {
      "post": {
        "tags": ["register"],
        "summary": "Register an agent wallet",
        "description": "Voluntarily register a wallet with optional metadata. Adds +15 identity points to the score and marks the wallet as registered on the leaderboard. GitHub URL is asynchronously verified against the GitHub API. Free — no payment required.",
        "operationId": "registerAgent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterBody" },
              "example": {
                "wallet": "0x3e4ef1f774857c69e33dddc471e110c7ac7bb528",
                "name": "DJD Score API",
                "description": "Reputation scoring for AI agents",
                "github_url": "https://github.com/owner/repo",
                "website_url": "https://djdagentscore.dev"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "First-time registration",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterResponse" } } }
          },
          "200": {
            "description": "Updated existing registration",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      },
      "get": {
        "tags": ["register"],
        "summary": "Look up agent registration",
        "description": "Returns registration metadata for a wallet. Free — no payment required.",
        "operationId": "getRegistration",
        "parameters": [
          { "$ref": "#/components/parameters/wallet" }
        ],
        "responses": {
          "200": {
            "description": "Registration found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "description": "Wallet not registered", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/v1/report": {
      "post": {
        "tags": ["report"],
        "summary": "Submit a fraud report",
        "description": "Submit a fraud or misconduct report against a wallet. Verified reports reduce the target wallet's composite score by up to 25 points. **Requires x402 payment of $0.02 USDC on Base.**",
        "operationId": "submitReport",
        "security": [{ "x402": [] }],
        "x-x402": { "price": "$0.02", "network": "base" },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["target", "reporter", "reason", "details"],
                "properties": {
                  "target":   { "type": "string", "pattern": "^0x[0-9a-fA-F]{40}$", "description": "Wallet being reported" },
                  "reporter": { "type": "string", "pattern": "^0x[0-9a-fA-F]{40}$", "description": "Wallet submitting the report (must differ from target)" },
                  "reason":   { "type": "string", "enum": ["failed_delivery", "payment_fraud", "impersonation", "malicious_behavior", "other"], "description": "Category of misconduct" },
                  "details":  { "type": "string", "maxLength": 1000, "description": "Supporting details and evidence (required, max 1000 chars)" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report submitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReportResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/v1/data/fraud/blacklist": {
      "get": {
        "tags": ["data"],
        "summary": "Fraud report check",
        "description": "Check whether a wallet has active fraud reports filed against it. Returns the count and most recent reports. **Requires x402 payment of $0.05 USDC on Base.**",
        "operationId": "getBlacklist",
        "parameters": [
          { "$ref": "#/components/parameters/wallet" }
        ],
        "security": [{ "x402": [] }],
        "x-x402": { "price": "$0.05", "network": "base" },
        "responses": {
          "200": {
            "description": "Blacklist status returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BlacklistResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/v1/data/economy": {
      "get": {
        "tags": ["data"],
        "summary": "Ecosystem economy metrics",
        "description": "Returns aggregated ecosystem health metrics: unique wallets scored, total queries, revenue, average scores, and more. Free — no payment required.",
        "operationId": "getEconomyMetrics",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "description": "Aggregation period (default: daily)",
            "schema": { "type": "string", "enum": ["daily", "weekly", "monthly"], "default": "daily" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of periods to return (max 90, default 30)",
            "schema": { "type": "integer", "minimum": 1, "maximum": 90, "default": 30 }
          }
        ],
        "responses": {
          "200": {
            "description": "Economy metrics returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "period": { "type": "string" },
                    "count": { "type": "integer" },
                    "metrics": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/v1/score/batch": {
      "post": {
        "tags": ["score"],
        "summary": "Batch score lookup",
        "description": "Score up to 20 wallets in a single request. Returns basic score data for each wallet. **Requires x402 payment of $0.50 USDC on Base.**",
        "operationId": "batchScore",
        "security": [{ "x402": [] }],
        "x-x402": { "price": "$0.50", "network": "base" },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["wallets"],
                "properties": {
                  "wallets": {
                    "type": "array",
                    "items": { "type": "string", "pattern": "^0x[0-9a-fA-F]{40}$" },
                    "minItems": 2,
                    "maxItems": 20,
                    "description": "Array of wallet addresses to score"
                  }
                }
              },
              "example": {
                "wallets": [
                  "0x3e4ef1f774857c69e33dddc471e110c7ac7bb528",
                  "0x1234567890123456789012345678901234567890"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch scores returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": { "type": "array", "items": { "$ref": "#/components/schemas/BasicScoreResponse" } },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "wallet": {
        "name": "wallet",
        "in": "query",
        "required": true,
        "description": "Base wallet address (EIP-55 checksummed or lowercase)",
        "schema": { "type": "string", "pattern": "^0x[0-9a-fA-F]{40}$" },
        "example": "0x3e4ef1f774857c69e33dddc471e110c7ac7bb528"
      },
...
    },
    "schemas": {
...
      "ReportResponse": {
        "type": "object",
        "properties": {
          "reportId":          { "type": "string", "format": "uuid" },
          "status":            { "type": "string", "enum": ["accepted"] },
          "targetCurrentScore":{ "type": "integer", "minimum": 0, "maximum": 100 },
          "penaltyApplied":    { "type": "integer" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request parameters",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "PaymentRequired": {
        "description": "x402 payment required. The response includes a WWW-Authenticate header with payment details. Use an x402-compatible client to pay and retry.",
        "headers": {
          "WWW-Authenticate": { "schema": { "type": "string" } }
        }
      },
...
    },
    "securitySchemes": {
      "x402": {
        "type": "http",
        "scheme": "x402",
        "description": "x402 micropayment protocol (EIP-3009 USDC transferWithAuthorization on Base). See https://www.x402.org for client libraries."
      },
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "Admin API key for management endpoints"
      },
      "bearerApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication (Bearer djd_live_...)"
      }
    }
  }
}